add svn lg command

+++ svn/cmd.sh
@@ -21,11 +21,11 @@ if [ -n "$(which svn 2>/dev/null)" ]; then
             diff-plain)
                 shift;
                 echo "$@"
-                $svn diff --diff-cmd diff "$@"
+                command svn diff --diff-cmd diff "$@"
                 ;;
             diff-color)
                 shift;
-                $svn diff --diff-cmd colordiff "$@"
+                command svn diff --diff-cmd colordiff "$@"
                 ;;
             diff-vim)
                 shift;
@@ -65,18 +65,18 @@ if [ -n "$(which svn 2>/dev/null)" ]; then
                 done;
 
                 local tmp=$(mktemp -t svn-diff-vim.XXXXXXXXXX || return 1).${file##*.}
-                $svn cat $rev "$file" > $tmp || return 1
+                command svn cat $rev "$file" > $tmp || return 1
                 chmod a-w $tmp || return 1
                 vimdiff "$file" $tmp || return 1
                 rm -f $tmp || return 1
                 ;;
             diff-filemerge)
                 shift;
-                $svn diff --diff-cmd $HOME/libexec/svndiff -x opendiff "$@"
+                command svn diff --diff-cmd $HOME/libexec/svndiff -x opendiff "$@"
                 ;;
             diff-less)
                 shift;
-                $svn diff --diff-cmd colordiff "$@" |less -FRX
+                command svn diff --diff-cmd colordiff "$@" |less -FRX
                 ;;
             pc|propcopy)
                 if [ $# -lt 4 ]; then
@@ -86,10 +86,13 @@ if [ -n "$(which svn 2>/dev/null)" ]; then
                 propName=$2
                 fromFile=$3
                 shift 3
-                $svn propset $propName "$($svn propget $propName $fromFile)" "$@"
+                command svn propset $propName "$(command svn propget $propName $fromFile)" "$@"
                 ;;
+            lg)
+                shift 1
+                command svn log "$@" | less -FX
             *)
-                $svn "$@"
+                command svn "$@"
                 ;;
         esac
     }
